--- Assets/VRM10/Editor/Components/SpringBone/VRM10SelectorWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96fdd3a9c7c5ca44b977ef564a2cf15b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
