+++ Assets/VRM10/Editor/Components/VRM10Window.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cc70b88276e055640aa0da2d937c8f39
+guid: d10db1ad6a9273e4698465a666033569
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
